% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_eq_binomial.R
\name{stamp_eq_binomial}
\alias{stamp_eq_binomial}
\title{stamp_eq_binomial}
\usage{
stamp_eq_binomial(x = 0, y = 0, ...)
}
\arguments{
\item{x}{a numeric indicating position on x axis}

\item{y}{a numeric indicating position on y axis}

\item{...}{}
}
\value{
a ggplot2 annotation layer which is the plotmath version of a function
}
\description{
Binomial Distribution in the  Probability  context. the binomial distribution provides a way to model the probability of observing a specific number of successes in a fixed number of trials, based on the underlying probability of success on each trial. This can be useful for making predictions, testing hypotheses, and assessing the reliability or efficiency of a process.
}
\examples{
library(ggplot2)
ggplot() +
   stamp_eq_binomial()
}
