% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_eq_ci.R
\name{stamp_eq_ci}
\alias{stamp_eq_ci}
\title{stamp_eq_ci}
\usage{
stamp_eq_ci(x = 0, y = 0, ...)
}
\arguments{
\item{x}{a numeric indicating position on x axis}

\item{y}{a numeric indicating position on y axis}

\item{...}{}
}
\value{
a ggplot2 annotation layer which is the plotmath version of a function
}
\description{
Confidence Interval in the  One-proportion z-test context. The confidence interval is a useful tool in hypothesis testing, as it allows us to estimate the range of possible values for the true population proportion, and can help us decide whether a null hypothesis is plausible or should be rejected.
}
\examples{
library(ggplot2)
ggplot() +
   stamp_eq_ci()
}
