% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_eq_ev.R
\name{stamp_eq_ev}
\alias{stamp_eq_ev}
\title{stamp_eq_ev}
\usage{
stamp_eq_ev(x = 0, y = 0, ...)
}
\arguments{
\item{x}{a numeric indicating position on x axis}

\item{y}{a numeric indicating position on y axis}

\item{...}{}
}
\value{
a ggplot2 annotation layer which is the plotmath version of a function
}
\description{
Expected Value in the  Probability Distributions context. The expected value of a probability distribution is a measure of the central tendency of the distribution, and represents the average or mean value that we would expect to observe if we were to repeat the random experiment many times. The expected value is a useful statistic because it provides a single number that summarizes the distribution of values that can be observed. It can be used to make predictions about future outcomes, as well as to assess the performance of statistical models or experimental designs.
}
\examples{
library(ggplot2)
ggplot() +
   stamp_eq_ev()
}
