% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_eq_geometric.R
\name{stamp_eq_geometric}
\alias{stamp_eq_geometric}
\title{stamp_eq_geometric}
\usage{
stamp_eq_geometric(x = 0, y = 0, ...)
}
\arguments{
\item{x}{a numeric indicating position on x axis}

\item{y}{a numeric indicating position on y axis}

\item{...}{}
}
\value{
a ggplot2 annotation layer which is the plotmath version of a function
}
\description{
Geometric Distribution in the  Probability  context. In each of these situations, the geometric distribution provides a way to model the probability of observing a specific number of trials until the first success occurs, based on the underlying probability of success on each trial. This can be useful for making predictions, designing experiments or systems, and assessing the reliability or efficiency of a process.
}
\examples{
library(ggplot2)
ggplot() +
   stamp_eq_geometric()
}
