% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_eq_normal.R
\name{stamp_eq_normal}
\alias{stamp_eq_normal}
\title{stamp_eq_normal}
\usage{
stamp_eq_normal(x = 0, y = 0, ...)
}
\arguments{
\item{x}{a numeric indicating position on x axis}

\item{y}{a numeric indicating position on y axis}

\item{...}{}
}
\value{
a ggplot2 annotation layer which is the plotmath version of a function
}
\description{
Normal Distribution in the  Probability  context. Normal distribution: The normal distribution is one of the most commonly used probability distributions in statistics, and is often used to model continuous variables that are approximately normally distributed. In R, the normal distribution is implemented using the dnorm(), pnorm(), qnorm(), and rnorm() functions.
}
\examples{
library(ggplot2)
ggplot() +
   stamp_eq_normal()
}
