% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/stamp_eq_sd_null_t_test.R
\name{stamp_eq_sd_null_t_test}
\alias{stamp_eq_sd_null_t_test}
\title{stamp_eq_sd_null_t_test}
\usage{
stamp_eq_sd_null_t_test(x = 0, y = 0, ...)

stamp_eq_sd_null_t_test(x = 0, y = 0, ...)
}
\arguments{
\item{x}{a numeric indicating position on x axis}

\item{y}{a numeric indicating position on y axis}

\item{...}{}
}
\value{
a ggplot2 stamp layer

a ggplot2 annotation layer which is the plotmath version of a function
}
\description{
Provides standard deviation equation for display on ggplot figure

Standard Deviation of the null distribution in the  One-sample t-test context.
}
\examples{
library(ggplot2)
   ggplot() +
   stamp_eq_sd_null_t_test()
library(ggplot2)
ggplot() +
   stamp_eq_sd_null_t_test()
}
