% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_eq_se_z_test.R
\name{stamp_eq_se_z_test}
\alias{stamp_eq_se_z_test}
\title{stamp_eq_se_z_test}
\usage{
stamp_eq_se_z_test(x = 0, y = 0, ...)
}
\arguments{
\item{x}{a numeric indicating position on x axis}

\item{y}{a numeric indicating position on y axis}

\item{...}{}
}
\value{
a ggplot2 annotation layer which is the plotmath version of a function
}
\description{
Standard error in the  One-proportion z-test context. The standard error (SE) of a one proportion z-test is a measure of the variability of the sample proportion estimate around the true population proportion. It tells us how much we expect the sample proportion to vary if we were to repeat the sampling process many times.
}
\examples{
library(ggplot2)
ggplot() +
   stamp_eq_se_z_test()
}
