% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stamp_eq_variance}
\alias{stamp_eq_variance}
\title{stamp_eq_variance}
\usage{
stamp_eq_variance(x = 0, y = 0, ...)
}
\arguments{
\item{x}{a numeric indicating position on x axis}

\item{y}{a numeric indicating position on y axis}

\item{...}{}
}
\value{
a ggplot2 stamp layer
}
\description{
Provides variance equation for display on ggplot figure
}
\examples{
library(ggplot2)
   ggplot() +
   stamp_eq_variance()
}
