% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_se_two_sample_t_test.R
\name{stamp_se_two_sample_t_test}
\alias{stamp_se_two_sample_t_test}
\title{stamp_se_two_sample_t_test}
\usage{
stamp_se_two_sample_t_test(x = 0, y = 0, ...)
}
\arguments{
\item{x}{a numeric indicating position on x axis}

\item{y}{a numeric indicating position on y axis}

\item{...}{}
}
\value{
a ggplot2 annotation layer which is the plotmath version of a function
}
\description{
Standard Error in the  Two-Sample T-test context.
}
\examples{
library(ggplot2)
ggplot() +
   stamp_se_two_sample_t_test()
}
