% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_statistic.R
\name{stamp_statistic}
\alias{stamp_statistic}
\title{stamp_statistic}
\usage{
stamp_statistic(x = 0, y = 0, ...)
}
\arguments{
\item{x}{a numeric indicating position on x axis}

\item{y}{a numeric indicating position on y axis}

\item{...}{}
}
\value{
a ggplot2 annotation layer which is the plotmath version of a function
}
\description{
Unstandardized statistic in the  One- proportion z-test context.
}
\examples{
library(ggplot2)
ggplot() +
   stamp_statistic()
}
