% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeset_eq_choose.R
\name{typeset_eq_choose}
\alias{typeset_eq_choose}
\title{typeset_eq_choose}
\usage{
typeset_eq_choose()
}
\value{
A character string containing a mathematical formula
}
\description{
The funtion_name equation for use in plain texts
}
\examples{
typeset_eq_choose
}
