% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeset_eq_ci.R
\name{typeset_eq_ci}
\alias{typeset_eq_ci}
\title{typeset_eq_ci}
\usage{
typeset_eq_ci()
}
\value{
A character string containing a mathematical formula
}
\description{
The funtion_name equation for use in plain texts
}
\examples{
typeset_eq_ci
}
