% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeset_eq_ev.R
\name{typeset_eq_ev}
\alias{typeset_eq_ev}
\title{typeset_eq_ev}
\usage{
typeset_eq_ev()
}
\value{
A character string containing a mathematical formula
}
\description{
The funtion_name equation for use in plain texts
}
\examples{
typeset_eq_ev
}
