% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeset_eq_geometric.R
\name{typeset_eq_geometric}
\alias{typeset_eq_geometric}
\title{typeset_eq_geometric}
\usage{
typeset_eq_geometric()
}
\value{
A character string containing a mathematical formula
}
\description{
The funtion_name equation for use in plain texts
}
\examples{
typeset_eq_geometric
}
