% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeset_eq_normal.R
\name{typeset_eq_normal}
\alias{typeset_eq_normal}
\title{typeset_eq_normal}
\usage{
typeset_eq_normal()
}
\value{
A character string containing a mathematical formula
}
\description{
The funtion_name equation for use in plain texts
}
\examples{
typeset_eq_normal
}
