% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeset_eq_sd.R
\name{typeset_eq_sd}
\alias{typeset_eq_sd}
\title{typeset_eq_sd}
\usage{
typeset_eq_sd()
}
\value{
A character string containing a mathematical formula
}
\description{
The funtion_name equation for use in plain texts
}
\examples{
typeset_eq_sd
}
