% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeset_eq_sd_null_t_test.R
\name{typeset_eq_sd_null_t_test}
\alias{typeset_eq_sd_null_t_test}
\title{typeset_eq_sd_null_t_test}
\usage{
typeset_eq_sd_null_t_test()
}
\value{
A character string containing a mathematical formula
}
\description{
The funtion_name equation for use in plain texts
}
\examples{
typeset_eq_sd_null_t_test
}
