% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeset_statistic.R
\name{typeset_statistic}
\alias{typeset_statistic}
\title{typeset_statistic}
\usage{
typeset_statistic()
}
\value{
A character string containing a mathematical formula
}
\description{
The funtion_name equation for use in plain texts
}
\examples{
typeset_statistic
}
